# aurora/foundry/server/swagger3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aurora",
    "description": "Produce predictions with the Aurora model",
    "version": "1.0.0"
  },
  "paths": {
    "/score": {
      "post": {
        "summary": "Create a new task",
        "operationId": "create_score_score_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Update an existing task",
        "operationId": "get_score_score_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the task",
              "examples": {
                "task_id": {
                  "value": "abc-123-def",
                  "summary": "Sample Task ID"
                }
              },
              "title": "Task Id"
            },
            "description": "The ID of the task"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Succeeds when the service is ready.",
        "operationId": "read_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/swagger.json": {
      "get": {
        "summary": "Get Swagger",
        "operationId": "get_swagger_swagger_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreationResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "CreationResponse",
        "example": {
          "task_id": "abc-123-def"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Submission": {
        "properties": {
          "data_folder_uri": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Data Folder Uri"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "num_steps": {
            "type": "integer",
            "title": "Num Steps"
          }
        },
        "type": "object",
        "required": [
          "data_folder_uri",
          "model_name",
          "num_steps"
        ],
        "title": "Submission",
        "example": {
          "data_folder_uri": "https://my.blob.core.windows.net/container/some/path?WRITABLE_SAS",
          "model_name": "aurora-0.25-small-pretrained",
          "num_steps": 5
        }
      },
      "TaskInfo": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "progress_percentage": {
            "type": "integer",
            "title": "Progress Percentage"
          },
          "success": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success"
          },
          "submitted": {
            "type": "boolean",
            "title": "Submitted"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "completed",
          "progress_percentage",
          "success",
          "submitted",
          "status"
        ],
        "title": "TaskInfo",
        "example": {
          "completed": true,
          "error_info": "Queued",
          "progress_percentage": 100,
          "submitted": true,
          "task_id": "abc-123-def"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Example 'Bearer myApiKey'",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}